test: move queue-split throughput tests to unit suite
The file used only fakeredis and threads (no ffmpeg), so the
integration mark violated the project's own placement rule and kept
three valuable topology tests out of the default run. The stage-stub
harness they shared with the DAG parallelism test moves to
tests/helpers/dag_harness.py instead of being imported across test
modules.
